+++ crates/bevy_reflect/derive/src/impls/mod.rs
@@ -3,17 +3,17 @@ mod common;
 mod enums;
 #[cfg(feature = "functions")]
 mod func;
+mod opaque;
 mod structs;
 mod tuple_structs;
 mod typed;
-mod values;
 
 pub(crate) use assertions::impl_assertions;
 pub(crate) use common::{common_partial_reflect_methods, impl_full_reflect};
 pub(crate) use enums::impl_enum;
 #[cfg(feature = "functions")]
 pub(crate) use func::impl_function_traits;
+pub(crate) use opaque::impl_opaque;
 pub(crate) use structs::impl_struct;
 pub(crate) use tuple_structs::impl_tuple_struct;
 pub(crate) use typed::{impl_type_path, impl_typed};
-pub(crate) use values::impl_opaque;
